+++ nflow-engine/src/main/java/com/nitorcreations/nflow/engine/service/WorkflowInstanceService.java
@@ -1,5 +1,6 @@
 package com.nitorcreations.nflow.engine.service;
 
+import static org.joda.time.DateTime.now;
 import static org.springframework.util.StringUtils.isEmpty;
 
 import java.util.Collection;
@@ -15,6 +16,7 @@ import com.nitorcreations.nflow.engine.workflow.definition.WorkflowDefinition;
 import com.nitorcreations.nflow.engine.workflow.instance.QueryWorkflowInstances;
 import com.nitorcreations.nflow.engine.workflow.instance.WorkflowInstance;
 import com.nitorcreations.nflow.engine.workflow.instance.WorkflowInstanceAction;
+import com.nitorcreations.nflow.engine.workflow.instance.WorkflowInstanceAction.WorkflowActionType;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
@@ -118,6 +120,27 @@ public class WorkflowInstanceService {
           .setStateText(instance.stateText).build();
     }
     workflowInstanceDao.updateWorkflowInstanceAfterExecution(instance, action);
+  }
+
+  /**
+   * Unschedule workflow instance in the database if it is not currently executing, and insert the workflow instance action
+   * if the actionDescription is not null.
+   *
+   * @param id The identifier of the workflow instance to be stopped.
+   * @param actionDescription The action description. Can be null.
+   * @param actionType The type of action.
+   * @return True if the workflow instance was stopped, false otherwise.
+   */
+  @Transactional
+  public boolean stopWorkflowInstance(int id, String actionDescription, WorkflowActionType actionType) {
+    boolean updated = workflowInstanceDao.stopNotRunningWorkflowInstance(id, actionDescription);
+    if (updated && actionDescription != null) {
+      String currentState = workflowInstanceDao.getWorkflowInstanceState(id);
+      WorkflowInstanceAction action = new WorkflowInstanceAction.Builder().setWorkflowInstanceId(id).setState(currentState)
+          .setStateText(actionDescription).setExecutionStart(now()).setExecutionEnd(now()).setType(actionType).build();
+      workflowInstanceDao.insertWorkflowInstanceAction(action);
+    }
+    return updated;
   }
 
   /**