+++ nflow-engine/src/main/resources/scripts/db/update-1.2.0-x/h2.update.ddl.sql
@@ -2,4 +2,4 @@ alter table nflow_workflow_definition add definition_sha1 varchar(40) not null d
 
 alter table nflow_workflow_action add type varchar(32) not null default 'stateExecution' check type in ('stateExecution', 'stateExecutionFailed', 'recovery', 'externalChange');
 
-alter table nflow_workflow add status varchar(32) not null default 'inProgress' check status in ('created', 'executing', 'inProgress', 'finished', 'manual');
+alter table nflow_workflow add status varchar(32) not null default 'inProgress' check status in ('created', 'executing', 'inProgress', 'finished', 'manual', 'stopped');